+++ tests/test_components/test_base.py
@@ -123,3 +123,12 @@ def test_equality():
     mnt2 = td.FluxMonitor(size=(1, 1, 0), freqs=np.array([1, 2, 3]) * 1e12, name="1")
 
     assert mnt1 == mnt2
+
+
+def test_special_characters_in_name():
+    """Test error if special characters are in a component's name."""
+    with pytest.raises(ValueError):
+        mnt = td.FluxMonitor(size=(1, 1, 0), freqs=np.array([1, 2, 3]) * 1e12, name="mnt/flux")
+
+    with pytest.raises(ValueError):
+        med = td.Medium(permittivity=1, name='"my med"')