+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+- Characters `"` and `/` not allowed in component names.
+
+### Fixed
+
+
 ## [2.6.1] - 2024-03-07
 
 ### Added